requirements.txt
python-dotenv
pandas
numpy
psycopg[binary,pool]
faker
tqdm